toolkit/components/passwordmgr/test/mochitest/test_autocomplete_highlight_username_only_form.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test form field autofill highlight</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content"></div>
<pre id="test">
Highlighting autofill fields in the test form

Fields that have been filled in automatically are highlighted by the pseudo class `:autofill`.
This test ensures that this pseudo-class is applied to filled-in fields and is also cleaned up afterwards.

<template id="form1-template">
  <form id="form1" action="https://autocomplete">
    <input type="text" name="uname" autocomplete="username">
    <button type="submit" name="submit">Submit</button>
  </form>
</template>

<script>
  const { ContentTaskUtils } = SpecialPowers.ChromeUtils.importESModule(
    "resource://testing-common/ContentTaskUtils.sys.mjs"
  );

  const formTemplate = document.getElementById("form1-template");

  add_setup(async () => {
    await setStoredLoginsDuringTest(
      [location.origin, "https://autocomplete", null, "user1", "pass1"],
      [location.origin, "https://autocomplete", null, "user2", "pass2"]
    );
    listenForUnexpectedPopupShown();
  });

  add_named_task("username field in username only form highlight on autocomplete", async () => {
    const form = setContentForTask(formTemplate);

    await openPopupOn(form.uname);
    synthesizeKey("KEY_ArrowDown");
    await synthesizeKey("KEY_Enter");

    const autofillResult = await formAutofillResult(form.id);
    is(autofillResult, "multiple_logins", "form has not been filled due to multiple logins");

    ok(form.uname.matches(":autofill"), "Highlight was successfully applied to the username field on username autocomplete");

    // Clear existing highlight on login fields. We check by pressing the tab key after backspace
    // (by shifting focus to the next element) because the tab key was known to cause a bug where the
    // highlight is applied once again. See Bug 1526522.
    form.uname.focus();
    synthesizeKey("KEY_Backspace");
    synthesizeKey("KEY_Tab");
    ok(!form.uname.matches(":autofill"),
       "Highlight was successfully removed on the username field");
  });
</script>
</body>
</html>
